backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
sqlalchemy==2.0.30
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==1.10.15
psycopg2-binary==2.9.9
email-validator==2.1.1